migration fixed

+++ database/migrations/2023_12_04_142938_create_posts_table.php
@@ -18,7 +18,8 @@ class CreatePostsTable extends Migration
 			$table->string('title');
 			$table->text('content');
             $table->timestamps();
-            $table->integer('user_id')->index();
+            //$table->integer('user_id')->index();
+            $table->unsignedBigInteger('user_id')->index();
             $table->foreign('user_id')->references('id')->on('users')
             //->onDelete('cascade')
             ;
